refactor(gcb): Move all GCB files into the same package
The GoolgeCloudBuildArtifact and GoolgeCloudBuildConfig classes
were in different packages; by moving these to the same package
as the other GCB classes we'll be able to reduce the visibility
of many GCB classes to package-private.
