+++ igor-web/src/main/java/com/netflix/spinnaker/igor/gcb/GoogleCloudBuildConfig.java
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package com.netflix.spinnaker.igor.config;
+package com.netflix.spinnaker.igor.gcb;
 
 import com.google.api.client.googleapis.javanet.GoogleNetHttpTransport;
 import com.google.api.client.http.HttpTransport;
 import com.netflix.spinnaker.igor.IgorConfigurationProperties;
-import com.netflix.spinnaker.igor.gcb.*;
+import com.netflix.spinnaker.igor.config.GoogleCloudBuildProperties;
 import com.netflix.spinnaker.igor.polling.LockService;
 import com.netflix.spinnaker.kork.jedis.RedisClientDelegate;
 import java.io.IOException;
